moved test scripts to a folder
